Adding flag for ability to fallback to screenshots if video capture is not supported.
